CDH-64723 FLUME-2957. Remove Guava from our public API
The only Guava class used in public API is the ImmutableMap in the

org.apache.flume.Context,
org.apache.flume.node.MaterializedConfiguration and
org.apache.flume.node.SimpleMaterializedConfiguration

classes.

This commit replaces these with java.util.Map.

Note: this is a breaking change as it modifies
the return type of a method on a public interface.

This closes #195

Reviewers: Ferenc Szabo, Mike Percy

(Denes Arvay via Ferenc Szabo)

Change-Id: Ibb449332773db351de703467e6c9abfcd881318f